combined shared utility files into one utilities.py
