--- secureli/modules/shared/utilities/secureli_meta.py
@@ -1,6 +1,0 @@
-from importlib.metadata import version
-
-
-def secureli_version() -> str:
-    """Leverage package resources to determine the current version of secureli"""
-    return version("secureli")
