+++ tests/modules/shared/utilities/test_git_meta.py
@@ -4,16 +4,14 @@ from unittest.mock import MagicMock
 import pytest
 from pytest_mock import MockerFixture
 
-from secureli.modules.shared.utilities import git_meta
+from secureli.modules.shared import utilities
 
 mock_git_origin_url = r"git@github.com:my-org/repo%20with%20spaces.git"
 
 
 @pytest.fixture()
 def mock_subprocess(mocker: MockerFixture) -> MagicMock:
-    mock_subprocess = mocker.patch(
-        "secureli.modules.shared.utilities.git_meta.subprocess"
-    )
+    mock_subprocess = mocker.patch("secureli.modules.shared.utilities.subprocess")
     mock_subprocess.run.return_value = CompletedProcess(
         args=[], returncode=0, stdout="great.engineer@slalom.com\n".encode("utf8")
     )
@@ -22,9 +20,7 @@ def mock_subprocess(mocker: MockerFixture) -> MagicMock:
 
 @pytest.fixture()
 def mock_configparser(mocker: MockerFixture) -> MagicMock:
-    mock_configparser = mocker.patch(
-        "secureli.modules.shared.utilities.git_meta.configparser"
-    )
+    mock_configparser = mocker.patch("secureli.modules.shared.utilities.configparser")
     mock_configparser_instance = MagicMock()
     mock_configparser_instance['remote "origin"'].get.return_value = (
         "https://fake-build.com/git/repo"
@@ -61,14 +57,14 @@ def mock_open_io_error(mocker: MockerFixture) -> MagicMock:
 
 
 def test_git_user_email_loads_user_email_via_git_subprocess(mock_subprocess: MagicMock):
-    result = git_meta.git_user_email()
+    result = utilities.git_user_email()
 
     mock_subprocess.run.assert_called_once()
     assert result == "great.engineer@slalom.com"  # note: without trailing newline
 
 
 def test_origin_url_parses_config_to_get_origin_url(mock_configparser: MagicMock):
-    result = git_meta.origin_url()
+    result = utilities.origin_url()
 
     mock_configparser.read.assert_called_once_with(".git/config")
     assert result == "https://fake-build.com/git/repo"
@@ -77,7 +73,7 @@ def test_origin_url_parses_config_to_get_origin_url(mock_configparser: MagicMock
 def test_current_branch_name_finds_ref_name_from_head_file(
     mock_open_git_head: MagicMock,
 ):
-    result = git_meta.current_branch_name()
+    result = utilities.current_branch_name()
 
     assert result == "feature/wicked-sick-branch"
 
@@ -85,10 +81,10 @@ def test_current_branch_name_finds_ref_name_from_head_file(
 def test_current_branch_name_yields_unknown_due_to_io_error(
     mock_open_io_error: MagicMock,
 ):
-    result = git_meta.current_branch_name()
+    result = utilities.current_branch_name()
 
     assert result == "UNKNOWN"
 
 
 def test_configparser_can_read_origin_url_with_percent(mock_open_git_origin: MagicMock):
-    assert git_meta.origin_url() == mock_git_origin_url
+    assert utilities.origin_url() == mock_git_origin_url
