+++ tests/modules/shared/utilities/test_usage_stats.py
@@ -4,7 +4,7 @@ from secureli.modules.shared.models.result import Result
 from secureli.modules.shared.models.scan import ScanFailure
 from secureli.repositories.repo_settings import TelemetrySettings
 from secureli.settings import Settings
-from secureli.modules.shared.utilities import usage_stats
+from secureli.modules.shared import utilities
 from unittest import mock
 from unittest.mock import Mock, patch
 
@@ -18,7 +18,7 @@ def test_that_convert_failures_to_failure_count_returns_correct_count():
         ScanFailure(id="testfailid2", file="testfile1", repo="testrepo1"),
     ]
 
-    result = usage_stats.convert_failures_to_failure_count(list_of_failure)
+    result = utilities.convert_failures_to_failure_count(list_of_failure)
 
     assert result["testfailid1"] == 2
     assert result["testfailid2"] == 1
@@ -27,7 +27,7 @@ def test_that_convert_failures_to_failure_count_returns_correct_count():
 def test_that_convert_failures_to_failure_count_returns_correctly_when_no_failure():
     list_of_failure = []
 
-    result = usage_stats.convert_failures_to_failure_count(list_of_failure)
+    result = utilities.convert_failures_to_failure_count(list_of_failure)
 
     assert result == {}
 
@@ -42,7 +42,7 @@ def test_that_convert_failures_to_failure_count_returns_correctly_when_no_failur
 )
 @patch("requests.post")
 def test_post_log_with_no_api_key(mock_requests):
-    result = usage_stats.post_log("testing", Settings())
+    result = utilities.post_log("testing", Settings())
 
     mock_requests.assert_not_called()
 
@@ -63,7 +63,7 @@ def test_post_log_with_no_api_key(mock_requests):
 )
 @patch("requests.post")
 def test_post_log_with_no_api_endpoint(mock_requests):
-    result = usage_stats.post_log(
+    result = utilities.post_log(
         "testing", Settings(telemetry=TelemetrySettings(api_url=None))
     )
 
@@ -87,7 +87,7 @@ def test_post_log_with_no_api_endpoint(mock_requests):
 def test_post_log_http_error(mock_requests):
     mock_requests.side_effect = Exception("test exception")
 
-    result = usage_stats.post_log("test_log_data", Settings())
+    result = utilities.post_log("test_log_data", Settings())
 
     mock_requests.assert_called_once_with(
         url="testendpoint", headers={"Api-Key": "testkey"}, data="test_log_data"
@@ -110,7 +110,7 @@ def test_post_log_http_error(mock_requests):
 def test_post_log_happy_path(mock_requests):
     mock_requests.return_value = Mock(status_code=202, text="sample-response")
 
-    result = usage_stats.post_log("test_log_data", Settings())
+    result = utilities.post_log("test_log_data", Settings())
 
     mock_requests.assert_called_once_with(
         url="testendpoint", headers={"Api-Key": "testkey"}, data="test_log_data"
@@ -132,7 +132,7 @@ def test_post_log_happy_path(mock_requests):
 def test_post_log_uses_settings_endpoint_if_no_env_endpoint(mock_requests):
     mock_requests.return_value = Mock(status_code=202, text="sample-response")
 
-    result = usage_stats.post_log(
+    result = utilities.post_log(
         "test_log_data", Settings(telemetry=TelemetrySettings(api_url="testendpoint"))
     )
